+++ dbinspector/src/main/kotlin/com/infinum/dbinspector/ui/databases/rename/RenameDatabaseState.kt
@@ -1,0 +1,8 @@
+package com.infinum.dbinspector.ui.databases.rename
+
+internal sealed class RenameDatabaseState {
+
+    data class Renamed(
+        val success: Boolean
+    ) : RenameDatabaseState()
+}