+++ dbinspector/src/main/kotlin/com/infinum/dbinspector/ui/databases/rename/RenameDatabaseViewModel.kt
@@ -1,4 +1,4 @@
-package com.infinum.dbinspector.ui.databases.edit
+package com.infinum.dbinspector.ui.databases.rename
 
 import android.content.Context
 import com.infinum.dbinspector.domain.UseCases
@@ -6,9 +6,9 @@ import com.infinum.dbinspector.domain.database.models.DatabaseDescriptor
 import com.infinum.dbinspector.domain.shared.models.parameters.DatabaseParameters
 import com.infinum.dbinspector.ui.shared.base.BaseViewModel
 
-internal class EditDatabaseViewModel(
+internal class RenameDatabaseViewModel(
     private val renameDatabase: UseCases.RenameDatabase
-) : BaseViewModel<Any, EditDatabaseEvent>() {
+) : BaseViewModel<RenameDatabaseState, Any>() {
 
     fun rename(
         context: Context,
@@ -26,7 +26,7 @@ internal class EditDatabaseViewModel(
                 )
             }
             if (result.isNotEmpty()) {
-                emitEvent(EditDatabaseEvent.Renamed(descriptor = result.first()))
+                setState(RenameDatabaseState.Renamed(success = result.firstOrNull()?.name.equals(newName)))
             }
         }
 }